# [Compatibility] Added INCRBYFLOAT command