--- a/libs/server/Storage/Functions/MainStore/RMWMethods.cs
+++ b/libs/server/Storage/Functions/MainStore/RMWMethods.cs
@@ -149,6 +149,14 @@ namespace Garnet.server
                         return false;
                     CopyUpdateNumber(-decrBy, ref value, ref output);
                     break;
+                case RespCommand.INCRBYFLOAT:
+                    value.UnmarkExtraMetadata();
+                    length = input.LengthWithoutMetadata - RespInputHeader.Size;
+                    // Check if input contains a valid number
+                    if (!IsValidDouble(length, inputPtr + RespInputHeader.Size, output.SpanByte.AsSpan(), out double incrByFloat))
+                        return false;
+                    CopyUpdateNumber(incrByFloat, ref value, ref output);
+                    break;
                 default:
                     value.UnmarkExtraMetadata();
 
@@ -320,6 +328,13 @@ namespace Garnet.server
                     if (!IsValidNumber(length, inputPtr + RespInputHeader.Size, output.SpanByte.AsSpan(), out var decrBy))
                         return true;
                     return TryInPlaceUpdateNumber(ref value, ref output, ref rmwInfo, ref recordInfo, input: -decrBy);
+
+                case RespCommand.INCRBYFLOAT:
+                    length = input.LengthWithoutMetadata - RespInputHeader.Size;
+                    // Check if input contains a valid number
+                    if (!IsValidDouble(length, inputPtr + RespInputHeader.Size, output.SpanByte.AsSpan(), out var incrByFloat))
+                        return true;
+                    return TryInPlaceUpdateNumber(ref value, ref output, ref rmwInfo, ref recordInfo, incrByFloat);
 
                 case RespCommand.SETBIT:
                     byte* i = inputPtr + RespInputHeader.Size;
@@ -606,6 +621,18 @@ namespace Garnet.server
                         break;
                     }
                     TryCopyUpdateNumber(ref oldValue, ref newValue, ref output, input: -decrBy);
+                    break;
+
+                case RespCommand.INCRBYFLOAT:
+                    length = input.LengthWithoutMetadata - RespInputHeader.Size;
+                    // Check if input contains a valid number
+                    if (!IsValidDouble(length, input.ToPointer() + RespInputHeader.Size, output.SpanByte.AsSpan(), out var incrByFloat))
+                    {
+                        // Move to tail of the log
+                        oldValue.CopyTo(ref newValue);
+                        break;
+                    }
+                    TryCopyUpdateNumber(ref oldValue, ref newValue, ref output, input: incrByFloat);
                     break;
 
                 case RespCommand.SETBIT: