--- a/playground/CommandInfoUpdater/SupportedCommand.cs
+++ b/playground/CommandInfoUpdater/SupportedCommand.cs
@@ -143,6 +143,7 @@ namespace CommandInfoUpdater
             new("HVALS", RespCommand.HVALS),
             new("INCR", RespCommand.INCR),
             new("INCRBY", RespCommand.INCRBY),
+            new("INCRBYFLOAT", RespCommand.INCRBYFLOAT),
             new("INFO", RespCommand.INFO),
             new("KEYS", RespCommand.KEYS),
             new("LASTSAVE", RespCommand.LASTSAVE),