Adding INCRBYFLOAT command to garnet

- [x] Add INCRBYFLOAT command
- [x] Add Test cases
- [x] Add documentation

Question:

Do you guys have any optimal way to `double` to ASCII` byte`?

Below is a benchmark with 2 approaches.

**DoubleToSpanWithMathAbsRound** => Same code as the PR
**DoubleToSpanWithDiyFp**  => Similar code as `string.Format("{0:0.####}")`. Below are the links to the code I used from .dotnet runtime repo
[Number.DiyFp.cs](https://github.com/dotnet/runtime/blob/5535e31a712343a63f5d7d796cd874e563e5ac14/src/libraries/System.Private.CoreLib/src/System/Number.DiyFp.cs)
[Number.Grisu3.cs - TryRunCounted](https://github.com/dotnet/runtime/blob/5535e31a712343a63f5d7d796cd874e563e5ac14/src/libraries/System.Private.CoreLib/src/System/Number.Grisu3.cs#L429)
**DoubleToSpanWithStringFormat** => Using `doubleValue.ToString("0.###############")` and using buffer copy to move the value to input span
**DoubleToSpanWithUtf8Formatter** => Using Utf8Formatter
**DoubleToSpanWithTryFormat** => Using `doubleValue.TryFormat(buffer, out var bytesWritten, "0.#################")`

| Method                                                 | Value                | Mean      | Error    | StdDev   | Gen0   | Allocated |
|------------------------------------------------------- |--------------------- |----------:|---------:|---------:|-------:|----------:|
| DoubleToSpanWithMathAbsRound                      | 10.23                |  32.27 ns | 0.399 ns | 0.333 ns |      - |         - |
| DoubleToSpanWithDiyFp                             | 10.23                |  51.40 ns | 0.619 ns | 0.579 ns |      - |         - |
| DoubleToSpanWithStringFormat                      | 10.23                | 121.94 ns | 1.474 ns | 1.307 ns | 0.0017 |      32 B |
| DoubleToSpanWithUtf8Formatter                     | 10.23                | 142.52 ns | 2.637 ns | 2.590 ns |      - |         - |
| DoubleToSpanWithTryFormat                         | 10.23                | 122.79 ns | 2.079 ns | 1.843 ns |      - |         - |
| DoubleToSpanWithMathAbsRound                      | 12.023642000155785   | 172.41 ns | 3.402 ns | 4.303 ns |      - |         - |
| DoubleToSpanWithDiyFp                             | 12.023642000155785   |  51.86 ns | 0.928 ns | 0.868 ns |      - |         - |
| DoubleToSpanWithStringFormat                      | 12.023642000155785   | 111.17 ns | 2.158 ns | 2.309 ns | 0.0029 |      56 B |
| DoubleToSpanWithUtf8Formatter                     | 12.023642000155785   | 188.86 ns | 3.793 ns | 3.895 ns |      - |         - |
| DoubleToSpanWithTryFormat                         | 12.023642000155785   | 121.57 ns | 1.504 ns | 2.825 ns |      - |         - |
| DoubleToSpanWithMathAbsRound                      | 85.98498198191919    | 161.47 ns | 1.592 ns | 1.330 ns |      - |         - |
| DoubleToSpanWithDiyFp                             | 85.98498198191919    |  50.77 ns | 0.834 ns | 0.740 ns |      - |         - |
| DoubleToSpanWithStringFormat                      | 85.98498198191919    | 107.39 ns | 1.761 ns | 1.647 ns | 0.0029 |      56 B |
| DoubleToSpanWithUtf8Formatter                     | 85.98498198191919    | 204.33 ns | 1.950 ns | 1.728 ns |      - |         - |
| DoubleToSpanWithTryFormat                         | 85.98498198191919    | 111.06 ns | 1.260 ns | 1.052 ns |      - |         - |
| DoubleToSpanWithMathAbsRound                      | 152.3645             |  59.95 ns | 0.985 ns | 0.922 ns |      - |         - |
| DoubleToSpanWithDiyFp                             | 152.3645             |  58.23 ns | 1.142 ns | 1.269 ns |      - |         - |
| DoubleToSpanWithStringFormat                      | 152.3645             | 128.90 ns | 2.206 ns | 1.956 ns | 0.0019 |      40 B |
| DoubleToSpanWithUtf8Formatter                     | 152.3645             | 220.70 ns | 4.394 ns | 5.557 ns |      - |         - |
| DoubleToSpanWithTryFormat                         | 152.3645             | 136.13 ns | 2.682 ns | 3.193 ns |      - |         - |
| DoubleToSpanWithMathAbsRound                      | 456621336.22         |  82.08 ns | 1.669 ns | 1.855 ns |      - |         - |
| DoubleToSpanWithDiyFp                             | 456621336.22         |  44.48 ns | 0.425 ns | 0.355 ns |      - |         - |
| DoubleToSpanWithStringFormat                      | 456621336.22         | 119.48 ns | 1.382 ns | 1.293 ns | 0.0024 |      48 B |
| DoubleToSpanWithUtf8Formatter                     | 456621336.22         | 248.25 ns | 4.419 ns | 4.134 ns |      - |         - |
| DoubleToSpanWithTryFormat                         | 456621336.22         | 114.32 ns | 2.104 ns | 1.968 ns |      - |         - |
| DoubleToSpanWithMathAbsRound                      | 9.874(...)2E+20 [21] | 249.52 ns | 4.870 ns | 5.980 ns |      - |         - |
| DoubleToSpanWithDiyFp                             | 9.874(...)2E+20 [21] |  37.10 ns | 0.761 ns | 0.712 ns |      - |         - |
| DoubleToSpanWithStringFormat                      | 9.874(...)2E+20 [21] | 126.49 ns | 1.706 ns | 1.513 ns | 0.0033 |      64 B |
| DoubleToSpanWithUtf8Formatter                     | 9.874(...)2E+20 [21] | 322.90 ns | 2.100 ns | 1.753 ns |      - |         - |
| DoubleToSpanWithTryFormat                         | 9.874(...)2E+20 [21] | 138.64 ns | 2.705 ns | 2.778 ns |      - |         - |

Even though the dotnet runtime approach does better in most cases in the benchmark, I have not done it as part of this PR because of 2 reasons, 1) it has lots of code and 2) for smaller numbers it always performs worse. If you guys like dotnet's approach, I can change it quickly.

Let me know if you guys have any other ideas, I can implement and compare the performance.