@badrishc 2024-10-04
Why not use Utf8Parser.TryParse? The open PR https://github.com/microsoft/garnet/pull/631 also uses it.
 https://github.com/dotnet/runtime/blob/main/src/libraries/System.Private.CoreLib/src/System/Buffers/Text/Utf8Parser/Utf8Parser.Float.cs#L63