@badrishc 2024-10-04
For other direction, check this:

https://learn.microsoft.com/en-us/dotnet/api/system.buffers.text.utf8formatter.tryformat?view=net-8.0