@badrishc 2024-10-04
might be best to just convert the double to the ROS<byte> with a sufficiently large buffer, to know how many bytes it uses. hopefully the conversion will not need to be repeated.