@badrishc 2024-10-08
@Vijay-Nirmal - there seems to be a perf regression with this PR, for Increment:

BDN paramters: `-f BDN.benchmark.Resp.RespParseStress.Increment`

main:

| Method    | Mean     | Error    | StdDev   | Allocated |
|---------- |---------:|---------:|---------:|----------:|
| Increment | 34.80 us | 0.354 us | 0.314 us |         - |

PR:

| Method    | Mean     | Error    | StdDev   | Allocated |
|---------- |---------:|---------:|---------:|----------:|
| Increment | 44.36 us | 0.842 us | 0.703 us |         - |

### Analysis

It seems this is related to the `stackalloc` of the larger sized buffer for doubles in the same method (NetworkIncrement). My suggestion is that we separate out NetworkIncrementByFloat into its own method.