@badrishc 2024-10-08
> @badrishc Interesting, For `Increment`, `isFloat` will always be false. Then how come the `stackalloc` is impacting?

I tried making `MaximumFormatDoubleLength` equal to `MaximumFormatInt64Length` and that seems to bring the perf back to closer to original levels, hence the suspicion.

> @badrishc Is there any thumb rule on when to add RespParseStress?

Ideally for all "fast" commands we add so that we can track future perf regressions. However it is understandably extra work, so it is mostly on a case-by-case basis for now.