@vazois 2024-10-09
> @badrishc Interesting, For `Increment`, `isFloat` will always be false. Then how come the `stackalloc` is impacting?
> 
> @badrishc Is there any thumb rule on when to add RespParseStress?
> 
> @vazois That a new things to me, I will add it, also same question Is there any thumb rule on when to add ClusterSlotVeficationTests?
> 
> I will do everything this weekend

I made some updates and additions in this PR (https://github.com/microsoft/garnet/pull/709) for slot verification.
If you are adding a multi key command you can check how MSET and MGET is being tested. Single key commands should be easier to understand by check SET and GET redirection tests.
You can also check the description of this PR (https://github.com/microsoft/garnet/pull/474) if you want to learn more about the redirection logic.