@badrishc 2024-10-09
> Can you please add a slot verification test in ClusterSlotVeficationTests?

We will merge this PR now. Vijay - if you can add the ClusterSlotVerificationTest in a separate PR, that would be great.